Add strict mode to renderer
