+++ src/lib/built-in/with.ts
@@ -6,6 +6,7 @@ const with_: Plugin = {
   filter: (binding) => binding.name === "with",
   async ssr({ binding, generated, value, bubble }) {
     let template: string | undefined;
+    const q = binding.quote;
 
     bubble(() => {
       if (value) {
@@ -14,7 +15,7 @@ const with_: Plugin = {
 
       let expr = "_ssr_with: { ";
       if (template) {
-        expr += `template: ${utils.quoteJsString(template, binding.quote)}, `;
+        expr += `template: ${q}${utils.escapeHtml(template)}${q}, `;
       }
       expr += `value: ${binding.expression} }`;
       generated.overwrite(...binding.range.offset, expr);
