+++ src/lib/exports.ts
@@ -1,5 +1,13 @@
-export * as parser from "./parser.js";
 export { default as builtins } from "./_built-ins.js";
+export {
+  BindingContext,
+  type BindingContextOptions,
+  type ChildContextOptions,
+} from "./binding-context.js";
+export { Binding } from "./binding.js";
+export { SSRError, SSRWarning } from "./diagnostic.js";
+export { evaluate, transform } from "./eval.js";
+export * from "./parser.js";
 export { type Plugin } from "./plugin.js";
-export { Binding, render, type SSROptions } from "./ssr.js";
+export * from "./renderer.js";
 export * as utils from "./utils.js";
