+++ tsconfig.node.json
@@ -7,7 +7,9 @@
   "compilerOptions": {
     "esModuleInterop": true,
     "allowSyntheticDefaultImports": true,
-    "types": ["node"]
+    "exactOptionalPropertyTypes": false,
+    "types": ["node"],
+    "outDir": "build"
   },
   "include": ["src"],
   "exclude": ["src/runtime"]
